Factor out the block hand templates to a standalone file
